minor tweaks, UC in seperate file
